[REFACTOR] Domain 레이어를 domain/attendance 모듈로 이동
- Entity 클래스들을 domain 모듈로 이동
- Repository 인터페이스를 domain 모듈로 이동
- 패키지 구조 정리 (org.sopt.official.domain.attendance)
