+++ domain/attendance/src/main/java/org/sopt/official/domain/attendance/entity/AttendanceButtonType.kt
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.sopt.official.domain.entity.attendance
+package org.sopt.official.domain.attendance.entity
 
 enum class AttendanceButtonType(
     val attendanceRound: AttendanceRound,
