+++ domain/attendance/src/main/java/org/sopt/official/domain/attendance/entity/AttendanceHistory.kt
@@ -1,6 +1,6 @@
 /*
  * MIT License
- * Copyright 2023-2024 SOPT - Shout Our Passion Together
+ * Copyright 2023 SOPT - Shout Our Passion Together
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -22,10 +22,10 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.sopt.official.domain.entity.attendance
+package org.sopt.official.domain.attendance.entity
 
-enum class EventAttribute {
-    SEMINAR,
-    EVENT,
-    ETC
-}
+data class AttendanceHistory(
+    val userInfo: AttendanceUserInfo,
+    val attendanceSummary: AttendanceSummary,
+    val attendanceLog: List<AttendanceLog>
+)
