+++ domain/attendance/src/main/java/org/sopt/official/domain/attendance/entity/AttendanceRound.kt
@@ -1,6 +1,6 @@
 /*
  * MIT License
- * Copyright 2023-2024 SOPT - Shout Our Passion Together
+ * Copyright 2023 SOPT - Shout Our Passion Together
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -22,16 +22,13 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.sopt.official.domain.repository.attendance
+package org.sopt.official.domain.attendance.entity
 
-import org.sopt.official.data.model.attendance.AttendanceCodeResponse
-import org.sopt.official.domain.entity.attendance.AttendanceHistory
-import org.sopt.official.domain.entity.attendance.AttendanceRound
-import org.sopt.official.domain.entity.attendance.SoptEvent
-
-interface AttendanceRepository {
-    suspend fun fetchSoptEvent(): Result<SoptEvent>
-    suspend fun fetchAttendanceHistory(): Result<AttendanceHistory>
-    suspend fun fetchAttendanceRound(lectureId: Long): Result<AttendanceRound>
-    suspend fun confirmAttendanceCode(subLectureId: Long, code: String): Result<AttendanceCodeResponse>
+data class AttendanceRound(
+    val id: Long,
+    val roundText: String,
+) {
+    companion object {
+        val ERROR = AttendanceRound(-2, "")
+    }
 }
