+++ domain/attendance/src/main/java/org/sopt/official/domain/attendance/entity/AttendanceUserInfo.kt
@@ -22,11 +22,11 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.sopt.official.domain.entity.attendance
+package org.sopt.official.domain.attendance.entity
 
-data class AttendanceLog(
-    val attribute: EventAttribute,
-    val attendanceState: String,
-    val eventName: String,
-    val date: String
+data class AttendanceUserInfo(
+    val generation: Int,
+    val partName: String,
+    val userName: String,
+    val attendancePoint: Number
 )
