+++ domain/attendance/src/main/java/org/sopt/official/domain/attendance/entity/EventAttribute.kt
@@ -1,6 +1,6 @@
 /*
  * MIT License
- * Copyright 2023 SOPT - Shout Our Passion Together
+ * Copyright 2023-2024 SOPT - Shout Our Passion Together
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -22,20 +22,10 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.sopt.official.domain.entity.attendance
+package org.sopt.official.domain.attendance.entity
 
-data class SoptEvent(
-    val id: Int,
-    val eventType: EventType,
-    val date: String,
-    val location: String,
-    val eventName: String,
-    val message: String,
-    val isAttendancePointAwardedEvent: Boolean,
-    val attendances: List<Attendance>
-) {
-    data class Attendance(
-        val status: AttendanceStatus,
-        val attendedAt: String
-    )
+enum class EventAttribute {
+    SEMINAR,
+    EVENT,
+    ETC
 }
