+++ domain/attendance/src/main/java/org/sopt/official/domain/attendance/entity/EventType.kt
@@ -1,6 +1,6 @@
 /*
  * MIT License
- * Copyright 2023 SOPT - Shout Our Passion Together
+ * Copyright 2023-2024 SOPT - Shout Our Passion Together
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -22,22 +22,10 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.sopt.official.domain.entity.attendance
+package org.sopt.official.domain.attendance.entity
 
-import org.sopt.official.data.model.attendance.AttendanceCodeResponse
-
-enum class AttendanceErrorCode(
-    val attendanceErrorCode: AttendanceCodeResponse,
-    val messages: List<String>
-) {
-    WRONG_CODE(AttendanceCodeResponse(-2), listOf("[LectureException] : 코드가 일치하지 않아요!")),
-    BEFORE_ATTENDANCE(AttendanceCodeResponse(-1), listOf("[LectureException] : 1차 출석 시작 전입니다", "[LectureException] : 2차 출석 시작 전입니다")),
-    AFTER_ATTENDANCE(
-        AttendanceCodeResponse(0),
-        listOf("[LectureException] : 1차 출석이 이미 종료되었습니다.", "[LectureException] : 2차 출석이 이미 종료되었습니다.")
-    );
-
-    companion object {
-        fun of(message: String) = entries.find { it.messages.contains(message) }?.attendanceErrorCode
-    }
+enum class EventType {
+    NO_SESSION,
+    HAS_ATTENDANCE,
+    NO_ATTENDANCE
 }
