+++ domain/attendance/src/main/java/org/sopt/official/domain/attendance/entity/SoptEvent.kt
@@ -22,11 +22,20 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.sopt.official.domain.entity.attendance
+package org.sopt.official.domain.attendance.entity
 
-data class AttendanceUserInfo(
-    val generation: Int,
-    val partName: String,
-    val userName: String,
-    val attendancePoint: Number
-)
+data class SoptEvent(
+    val id: Int,
+    val eventType: EventType,
+    val date: String,
+    val location: String,
+    val eventName: String,
+    val message: String,
+    val isAttendancePointAwardedEvent: Boolean,
+    val attendances: List<Attendance>
+) {
+    data class Attendance(
+        val status: AttendanceStatus,
+        val attendedAt: String
+    )
+}
