+++ domain/attendance/src/main/java/org/sopt/official/domain/attendance/repository/AttendanceRepository.kt
@@ -1,0 +1,37 @@
+/*
+ * MIT License
+ * Copyright 2023-2024 SOPT - Shout Our Passion Together
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+package org.sopt.official.domain.attendance.repository
+
+import org.sopt.official.domain.attendance.entity.AttendanceCodeResponse
+import org.sopt.official.domain.attendance.entity.AttendanceHistory
+import org.sopt.official.domain.attendance.entity.AttendanceRound
+import org.sopt.official.domain.attendance.entity.SoptEvent
+
+interface AttendanceRepository {
+    suspend fun fetchSoptEvent(): Result<SoptEvent>
+    suspend fun fetchAttendanceHistory(): Result<AttendanceHistory>
+    suspend fun fetchAttendanceRound(lectureId: Long): Result<AttendanceRound>
+    suspend fun confirmAttendanceCode(subLectureId: Long, code: String): Result<AttendanceCodeResponse>
+}
